added missing deps
created new memcahced module

resolved nova authorized_keys file distribution issues

fixed remaining horizon issues and inconsistencies

removed horizon from openstack common, as its no longer needed.

adjusted to host name check to ensure that we only ever have a host name that is NO longer than 63

I guess I need more sleep, i missed my temp code for testing

made container name shortening smarter

removed container name munging

We were doing container name munging when the length of the container name was > 63. This however was
causing containers to not be in the correct host groups. To avoid general issues the munging code was
removed and replaced with a system exit stating that the hostnames are too long. While this is not
ideal it is the best way to ensure that the system is built correctly 100% of the time.

updated support plays to use new key distribution

fixed typos