Glance configuration changes
This change makes notification_driver configurable and only writes
messaging configuration options to glance-api.conf if
notification_driver is set to 'messaging' (it defaults to 'noop').

We also set a default value for 'swift_store_endpoint_type' so that
an ansible run will not bomb if the user does not have
'glance_swift_store_endpoint_type' defined in their user_variables.yml
file.