Adjusted pip install config and repos
This PR ensures that the environment is always pointing at a frozen repository of PIP packages which consists of prebuilt wheels. A script was created to allow for packages and apps to be easily updated.

Resolves:
* https://github.com/rcbops/ansible-lxc-rpc/issues/113
* https://github.com/rcbops/ansible-lxc-rpc/issues/120

Begins resolving:
* https://github.com/rcbops/ansible-lxc-rpc/issues/31
* https://github.com/rcbops/ansible-lxc-rpc/issues/53

set all project group_vars to a given sha

added retry

added tempest