Pin Glance to Juno master and update configs
Move all Swift info to the bottom of the DEFAULT options section

Because there is now a glance_store section, having that in the middle of the
swift store conditional caused the filesystem options to be moved out of
the DEFAULT scope. This makes sure everything is in the right section.