Idempotency for LXC cloning
Previously, clone operations where the new container already exists were
reporting as failed. This commit modifies that behavior so that if the
new container exists, it returns the facts for that container instead of
failing.

To facilitate these changes, a _container_exists method was added which
returns True or False based on whether or not the container name is in
the output from self._list().

I also made a few fixes in the documentation by removing trailing spaces
and fixing a typo in the state on line 222.

(cherry picked from commit 7032f4bb1ee4a3012084ef2cce6b543761f0a137)