Adds a simple to use and cron script for python packages
This script is intended as a simple method for croning the
creation of python wheels.

updated wheel builder python app

This commit increases performance when building all of the python packages
that our systems depended on. This also provides the system a capability
to build specific versions of packages that may be required and or duplicated
throughout the various projects.

IE a particular project may require a specific version of a package to be
installed that may be a lower revision that what is already built.

This resolves Issue: https://github.com/rcbops/ansible-lxc-rpc/issues/252

Added the ability to store git repos locally

This Commit provides a function to clone all of the git repos in a
given project to some local directory.

(cherry picked from commit 00d2834c3ada9349d15dec7355a25943479ffd55)