Be more specific w/ cluster check
I built a 3 controller setup and noticed that rabbitmq was not
clustering correctly.  Upon closer inspection, the issue was due to the
default cluster name of 'rpc' being found in a cluster_status due to
my hosts have a hostname of rpc-controller-X.  This change is more
specific in what we search for in the cluster_status output so we do
not get false positive backs from the cluster name being reflected in
the container hostnames.

Closes #256

(cherry picked from commit f9b4d30186ee0c4bfe075245978264872c93344f)