Default to creating agent token for each entity
Currently, we require the operator to specify an existing agent token
to use for all entities.  This means the operator needs to make a call
to MaaS to determine the agent token ID before being able to deploy the
environment.  This change creates a new agent token w/ label matching
the entity in question and configures the agent to use that token.  If
the user variable maas_agent_token is defined, then we skip creating an
agent token and use the token specified.

Closes issue #263

(cherry picked from commit 1752537874c60fb66bba020101993205c1c755df)