Check md5sum of swift ring .builder files before adjusting ring.
For safety - if the ring on the remote hosts is not empty, or the same as the local
host, we should abort and ask the user to check the consistency of the rings.

This will prevent a situation where the rings are inconsistent, but still allow new hosts
to be added to the ring. E.g. if the rings are removed from the local host for some reason,
this will cause the run to fail. Or if the rings aren't moved to the nodes in a uniform way.

Allows the user to manually intervene and fix the inconsistencies.