This commit allows for more galera/mysql tuning
The change allows the user to change the "mysql_max_connections" as variable
If unset, the default will be (100 * ansible_processor_vcpus). This is based
on the facts gathered by ansible. All galera nodes will be used in computing
the value of the max_connections and the lowest denominator will be used.

This change also modifies the count of wsrep sync thread based on the
the number of CPU cores available on a given host as defined by ansible
facts. This function can also be overridden using the "galera_wsrep_slave_threads"
variable.