Updated cinder.conf to allow for AZ setup
This commit provides vribales to all for the creation of cinder
availability zones. With this setup you can now create a cinder
volume container in a targetted availability zone and define a
defaul az. The new variables are:

* cinder_storage_availability_zone
* cinder_default_availability_zone

Each variable is a string value. The variable **cinder_default_availability_zone**
is optional and will default to the the scopped **cinder_storage_availability_zone**
variable when undefined per the documentation in Openstack ``cinder.conf``.

These changes were documented in the example configuration files as well.