Changed the container interaction process
This changes the way that containers are interacted with. With this
change, container actions are deletgated to the host instead of looping
through the hacky mess that we were doing. This change will make it
so that the entire container process is faster.

This also removes the needs for the "/openstack/monitoring" directory which
was held over cruft from long ago. This should address the race condition
when delegating to a host and the monitoring directory attempts to be created
at the same time on the same host.

Closes-Bug: #1399427
Change-Id: Ifaa0fa5719f79180610b4a63d590ca8bc681f87d
(cherry picked from commit fea671ec1692617d0c64c8b3d27df530bf58c004)