Change nova scheduler filters
This commit changes AggregateCoreFilter to CoreFilter and
AggregateDiskFilter to DiskFilter.  Whilst these filters work even if
aggregates are not being used, icehouse does not have
AggregateDiskFilter so there CoreFilter and DiskFilter need to be used.
This change is more to bring the two branches in line, and if anyone
does explicitly need the aggregate filter then that can be added via a
local override.

Change-Id: I6cc01139db7274f80d138020c97368a0faece771
Closes-Bug: 1399366