Allow container directories to have non-standard modes
This change adds the ability to customise the mode applied to a
directory in the container_directories group_vars.  Additionally, we
change the default mode of a directory with an unspecified mode to 700.
This latter change requires us to flip the following back to 755
otherwise the respective service throws a permission denied error:

* /var/www/cgi-bin/keystone
* /var/lib/neutron
* /var/lib/nova

A fair bit of additional testing will need to be done before this
merges.

Change-Id: I40b0052876078c3472e88594a5c0d7e039b9817a
Closes-Bug: 1399428
(cherry picked from commit 1978bf5fbb6c6df435ba513220e2baa33e065848)