Correct interface information for bonds
When generating bonded interfaces, we still want to have an IP available
if the user has defined it. This change will make sure that the address,
netmask, and DNS information gets set on bonded interfaces.

Also, tagged interfaces that are based off a bond need to be set to
`manual`, so that they are not forced to have IPs.

Closes-Bug: #1402028
Change-Id: I6ad29f875be6ae8b770f0608874f6a650204f681