Correct Swift service health checks
The account, container, and object service checks currently fail
because they are trying to connect to the {{ ansible_ssh_host }}
address. The services are only bound to the storage network
interface. This does not effect the proxy service as it is bound
to all interfaces.

Additionally, the health check url is misspelled for all services.

  - Use the storage network address for Swift service checks
  - Correct the health check url spelling

Change-Id: I581ab1e58464e3a99edeb52039a7b8f6e88f3149
Closes-Bug: #1401701