Add swift to the all-in-one build check
This patch adds swift to the all-in-one build check used for testing
and gate checks. The script will deploy swift and configure glance to
use swift by default, but this can be disabled by setting the
environment variable DEPLOY_SWIFT to be any value other than 'yes'.

This patch also adds comments to various parts of the check script for
clarity, removes br-snet as it is unused, and has a sprinkle of
refactoring for improved health and happiness.

Change-Id: Iabce466018d3895f61dc736087a14b15af08a35b
Closes-Bug: #1401287