Add swift_all inventory to swift-common play
This patch adds the swift_all inventory to the swift-common play to
ensure that the correct service name is used when the play tries to
install the openstack software. Without including the inventory a
different service name is used when the software installation is
executed, resulting in the play attempting to install neutron instead.

Change-Id: Ie15f96ef22ff18fa3b4bc145174d6ce21e9fd69a
Closes-Bug: #1402582