Add swift_all inventory to swift account/container/object plays
This patch adds the swift_all inventory to the swift account, container
and object plays to ensure that the correct service name is used when
the play configures the upstart services.
Without including the inventory a different service name is used,
resulting in upstart trying to find the config files in /etc/neutron.

Change-Id: I725e87fca647a21ea0e2d64952626b6f6fc6e2a3
Closes-Bug: #1402617