Do not recursively chmod/chown files in container_directories
Currently, we recursively chmod/chown files present in the directory
being iterated over in the contain_directories variable.  This is fine
on first run as most of these directories (if not all) do not exist and
are therefore empty.  However, when the container_common role is
re-applied, all files within will be altered.  According to
http://docs.ansible.com/file_module.html, non-existent parent
directories will get automatically created when state=directory,
recurse=yes is not necessary.  This is presumably why recurse=yes was
originally specified.

This commit also explicitly creates /var/lib/nova/cache to ensure it's
created with the right permission/ownership.  Lastly, we set some valid
ownership/permissions on /var/lib/nova/instances.  This was implicitly
happening with recurse=yes as the directory gets created in another
role prior to container_os_setup.yml being run.

Change-Id: Ibe54e19daa7e6d71dbde0f1ac59d86fc31cce7e5
Closes-Bug: #1403917