Ensure that permissions are permissible enough for the rsyslog container
The rsyslog container is unable to read the log files produced by the
other containers due to permissions issues and inconsistant user:group
ownership and user:group presence between containers.

The simplest way of dealing with this is to ensure that the log
directory allows read and traversal rights for all users.

This patch ensures that the log directory is always created by the
container_common role for containers.

This patch also add the processing of swift logs to logstash. The
template was always there, but had no task which deployed it.

Change-Id: I6ec12b2f2302e0cffc2a36096d38cb85e4b3ab04
Closes-Bug: #1407633
Related-Bug: #1403890