Implement logstash filter changes to improve log processing
This patch implements changes to the logstash filter configurations in
order to improve log processing such that all logs are tagged and passed
through to elasticsearch properly. It implements a change in the initial
'rsyslog' filter logic and several tweaks in the follow-up filters,
resulting in no more grok parse failures.

This patch also includes log parsing for swift, neutron and the mysql
slow query log.

Co-Authored-By: Steve Lewis <steve.lewis@rackspace.com>
Change-Id: Id6180b080f11fada00386e39c64f8d7338f8c94d
Closes-Bug: #1403890
Closes-Bug: #1399371
Closes-Bug: #1399367