add remote maas check for ssl cert expiry
if setting up your external endpoints to be behind an ssl cert on your
load balancer, it might be quite useful to get an alert when the
cert is due to expire.  This commit adds that check in a new
playbook that would need to be run manually if you are setting up
this way.

Closes-Bug: 1408335

Change-Id: Icc06e4c8e00b44df7540bb8d33f0d0afba20e4f6
(cherry picked from commit 6276950833189f227431181fb6a1d39b5e428fde)