Add socat to galera group_vars
The icehouse branch is structured a bit differently and with the
previous change applied socat is still not getting installed.  This
current change adds socat to galera group_vars which should result in it
getting installed correctly.

On a side note, it should be safe to remove
rpc_deployment/vars/repo_packages/galera.yml since that file is not
getting included anywhere, however I'd rather not make any large
changes to the icehouse branch at this time.

Change-Id: Ia99f5ff7ca5ebf107f971d4e8a159fa1ecafdba4
Closes-Bug: #1408138