Adjust os-ansible-aio-check to work for all branches
The icehouse branch does not currently have support for using the conf.d
directory for additional user configuration. This patch implements all
the configuration, except swift (which is not available in the icehouse
branch) into a monolothic user configuration file. This ensures
compatibility of the configuration process for all branches.
This patch also ensures that the plays executed are appropriate for the
branches - the icehouse branch has different names for some of the
plays.

Change-Id: Iffad2ffce93e8173da68863210afba0aa105622d
Closes-Bug: #1413606
(cherry picked from commit 8a21f096577a20be1aafdb2e3656cf5b06ff2f90)