Allow setup of specific MaaS filesystem monitors
By default will generate a list of filesystems that are ext or xfs and
set these up to be monitored with a specified threshold set in
user_variables.

Can loop through list of filesystems/thresholds specifically set in
user_variables if needed, and the default for a server can be overridden
in the rpc_user_config.yml per host.

The user_variables.yml sample file, and the rpc_user_config.yml files
have been updated to reflect these changes.

Change-Id: I1959a630e2c603a76001f52db6b027bf71124c54
Closes-Bug: #1414249