Swift dedicated replication init script fix
Add a check to ensure that the dedicated server-replicator init script
is only added if a dedicated replication network is in use.

This resolves an issue where the syslog error logs are filled with
attempts to restart an unrequired service.

Change-Id: I068608e128e1a076fe1cf7669eb80e0e2b7ebea6
Closes-Bug: #1414705
(cherry picked from commit ac50615faef2d453663991a23f0abd33e8210f24)