Set the binlog expire option in Galera
This option was left unset which can cause the bin log to grow forever.
Because we generally run in containers backed by LVM and the space which
the container is allowed to grow is limited. With that in mind we need to
to do everything possible to ensure that we do not create situations
that result in 100% full FS. This change sets the ``expire-logs-days``
to "7" which should be enough to provide for the optional bin log along
while also keeping the FS tidy.

Change-Id: I2c89d85ec77122a54419bdb2d5790e2a6e49fed8
Closes-bug: 1402487