Pinned Oslo Messaging version to version 1.4.1
This change pins oslo messaging to version 1.4.1 which will satisfy
both icehouse and juno requirements. This change was in direct
response a support request.

The issue with the latest Oslo Messaging for Juno, Oslo 1.5.1,
is that there is a potential for it to open too many connections
to rabbitmq. While this behavior is present in version 1.4.1 as well
it is less so than in 1.5.1. At this time pinning the version will
create a better overall deployment experience.

Change-Id: I37793557d53fc5965547d2b8cc3ca44812def8db
Partial-Bug: 1406629