Allow the pinning of apt package versions
This patch implements the capability to pin package versions using apt
preferences. This prevents the unwitting upgrading of packages when
doing general packages deployment and maintenance.

The list of packages and versions can be overridden using user
variables.

For each item in 'apt_pinned_packages'
 - package: package name
 - version: package version to pin
 - priority: pin priority (default is 900, over 1000 will make apt force
   a downgrade)

The patch also removes the usage of the LXC PPA - it was only deployed
if the operating system was < 14.04 and the reality is that the entire
project has not been tested on a platform < 14.04. The removal of the
PPA is thus to reduce confusion.

It also includes a known set of packages which are important to pin for
a stable OpenStack deployment. The packages pinned use sources and
version which should be available for a typical 'Long Term Support'
time-frame.

Closes-Bug: #1412898
Change-Id: I747e233a2c2d7e2c30e4d653cbeaca2e3cbe249a
(cherry picked from commit f2ef204f97f0811c23169baf131546699b7a1a1b)