Adds basic host kernel optimisation
This change adds three values to basic kernel optimisation which
should result in a more stable system.

Values that will be added in:
  - { key: 'vm.dirty_background_ratio', value: 5 }
  - { key: 'vm.dirty_ratio', value: 10 }
  - { key: 'vm.swappiness', value: 5 }

These values should improve general system operation and performance
when hosts have a potential for high IO. This is especially true in
a containerized environment when there are more than 20+ containers
running on a single host all potentially using LVM as the backend.

Change-Id: Ie4bf8d4d7e1b77c12bb3abec1b2d75c47ed29895
Closes-Bug: 1409755