Remove quotes to prevent duplicate alarms
The memory alarm has a space in the name. This is causing it to get
surrounded by single quotes, which in turn means grep finds no matches.

This patch removes the quotes to prevent the issue. There may be a more
general issue to address with quotes but it is outside the scope of
this patch.

Change-Id: I5ce02ee8499e593f5813ffe277d3869e7c28d691
Closes-bug: #1416443