Add ssh_retry connection plugin
The default ssh connection plugin will cause a task to fail if a
connection cannot be made first time. The failures have been found to
cause a number of builds to fail.

This patch adds a new connection plugin called ssh_retry and sets it as
the default one to use.

The plugin can be enabled by setting the following options in
ansible.cfg:

    [defaults]
    connection_plugins = plugins/connection_plugins
    transport = ssh_retry

    [ssh_retry]
    retries = 3

Note, the default retries is 3.

Change-Id: Ic187fb154cfa7b6fa95b19bee4757ec976f3f368
Co-Authored-By: Hugh Saunders <hugh@wherenow.org>
Closes-Bug: #1404343