Remove ImportError handling in dynamic_inventory
The import of netaddr any yaml is done within a try/except block. If an
ImportError occurs an error message is printed and the script ignores
it. Ansible does not display the printed message and the script will
still fail, just later on.

This commit removes the exception handling so that if there is an
ImportError the script will fail straight away.

Change-Id: Iff935db62bde954ac4ff0b920409c18dd494bdff
Closes-Bug: #1420312