Retry installing apt packages by default
There are a number of tasks used to install apt packages. Some of the
tasks will retry on failure and others will not. These tasks are
susceptible to failure and so it makes sense that they retry if the
task fails.

This commit updates the apt tasks so that they all retry with the
default retries and delay values of 3 and 5 respectively. In addition
all the tasks will attempt to update the cache if it is older than
600 seconds.

Some of the apt tasks had the state set to latest, this has been
changed to present because upgrades should be handled by specific
upgrade tasks.

Closes-Bug: #1421238
Change-Id: I8c20c5c0a2ce87706a99734adcdf55c93efb405c