Add retry and fallback server capability to apt-key tasks
This patch adds retries (with a delay) to the common and
container_common apt-key tasks, then also includes an optional retry
using a fallback keyserver for the appropriate tasks.

The patch also removes the conditional for the platform version as it is
not entirely relevant at this stage - the whole environment is only
tested to work on Ubuntu 14.04 anyway.

Closes-Bug: #1423123
Change-Id: Ica590a61319f60beaa8c6d6f4b8210400efca922