Updated contributing guidelines
This change updates the contributing guidelines per the meeting discussion
from earlier today. The contributing guidelines have been work on via the
community meeting forum and been adjusted to meet this projects goals.
The amendments to this file have been based on the content found here:
https://etherpad.openstack.org/p/ansible-contrib which was peer reviewed
by members of the `os-ansible-deployment` community.

Change-Id: I66cf4314a1871c396c1deee62342a39f5c6430de