Fail when no tests are run
Currently when testr fails to list any tests, no tests are run, so none
fail so the tempest script exits with success.

This patch adds some error handling around testr list-tests so that
errors are output to the console and the script halts on error.

There are also two tests added to check that full_test_list and
test_list are not empty.

Change-Id: Ieeb5957c9e47489a90a29c6f5aef166ad4e8c11e
Closes-Bug: #1425318