Add flat network gateway to br-vlan for AIO builds
When an AIO is built using the included scripts and networking config,
the public network has a gateway IP of 172.29.248.1 from the
172.29.248.0/22 network, which is not expected or configured to exist
anywhere. This can cause issues when using floats or in general if
some communication is desired which uses the public side gateway of
the neutron-routed network.

A simple solution is to simply drop 172.29.248.1/22 on br-vlan via the
interface config, which allows the traffic to do the needful, at least
as far as tempest's requirements are concerned. In modern
Debian/Ubuntu, this can be accomplished with another "iface" stanza
with its own "address" directive to add the additional address.

Change-Id: I79897bc4e4d7eb7d55ad3c12f55a339dfef869e1
Closes-Bug: #1425717
Related-Bug: #1425255