Fix co-dependent tempest issues
We noticed at the gate for icehouse alone that tempest tests were not
being collected or run. This was due to the fact that the version of
tempest we checkout has vendorized oslo.log code from oslo-incubator
which has a dependency on oslo.serialization.

We don't deploy swift in icehouse so tempest tests are not going to
work.

Co-Authored-By: Hugh Saunders<hugh@wherenow.org>

Change-Id: I416c4db4c6689cdcd701747e5cb1d566eae72140
Closes-bug: 1425767