Symlink /openstack/log so gate logs can be published
Ant needs a path that is local to the directory from which it operates.
To be able to publish our logs, we need to symlink /openstack/log to
/home/jenkins/workspace/os-ansible-deployment-dvsm-check-commit/logs.

This also allows us to remove the log publisher we wrote and just use
the devstack-logs publisher instead.

Conflicts:
	scripts/gate-check-commit.sh

Related-Bug: 1425482
Depends-On: If94e6b97c159000c83fadc6e4dd72d56ee90d3b9
Change-Id: Iee055beb853cbb76a073c8060d6920d317f5bd64
(cherry picked from commit 4d675bf7f576d4c63513f4f28ac713ffe21fb8e2)