Change testr to run in serial
This change implements the change to allow testr to run in serial which
should resolve most issues with tempest on a resource constrained host.

The problem that is being resolved here is that testr will, by default,
run in multi-threaded and has a relitively high probability to result in
a 503. This primarily impacts the gate jobs as they are built on a hosts
with 4 CPUs and 8GB of ram.

By default if the tempest script is run, testr will run multi-threaded.
However, if the environment variable RUN_TEMPEST_OPTS is set to
'--serial' then it will run in serial. This is done for the
gate-check-commit script so that the change only affects the AIO gate
check.

An additional environment variable TESTR_OPTS is added for further
flexibility.

Conflicts:
        scripts/gate-check-commit.sh
        scripts/run-tempest.sh

The above conflicts were due to the difference between master and this
branch with regards to the environment variable. In master all variables
are exported. This difference will be addressed in a subsequent patch.

Co-Authored-By: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>
Closes-Bug: #1428847
Change-Id: I8c7ac72ef46ae3f3af3d3503081bec0cbe7276e9
(cherry picked from commit 255b94d425c0e6bb6a78868b12e5ef1fb4346354)