Trim ansible's output to console.log
Since I2786f73334b99bfa98d344c50337c86f75fae273 ansible's output is
logged properly to a file and published at the end of a gate check.
This trims the output from ansible to the console to make the console
log that is published far shorter and easier to read.

Co-Authored-By: Ian Cordasco <ian.cordasco@rackspace.com>
Change-Id: I98fb57a56c1b7ddd6558e265cf6e65b159a05123
Partial-bug: 1425482