Set backend_alg for swift_proxy to source
This is a backport of 94ba3bc and 51abd0b from master to juno branch.
Commit 94ba3bc set HAProxy balance algorithm to "source" for swift-proxy
(required by Tempest which makes a lot of quick successive calls to
swift-proxy and were failing when using the default algorithm of
"leastconn").  Commit 51abd0b went back and cleaned up all the
incorrectly referenced variables (_backend_alg to _balance_alg) as the
change made in 94ba3bc wasn't actually taking effect.

Note that the ansible variable name in master is haproxy_balance_alg
whilst in juno it is hap_balance_alg.

Change-Id: I993e019c06f6d7ea7f77dc741d3d418e78ce2ca9
Closes-Bug: #1428081
Closes-Bug: #1429109