Fix gate: tempest and galera updates
Update tempest pin:
Tempest was installing v3 which is out of step with the other branches
and does not include the necessary patches required for networking tests
to work as we expect it to. This change installs a specific sha that is
newer than v3 and in line with the other branches. The master branch was
not selected to pin against as it's a moving target and disruptive to
our development process.

Pin tempest-lib:
This patch also pins tempest-lib to 0.2.1 as a later change in its
requirements was introduced which breaks the tempest installation.

Change galera package name:
The mariadb repository appears to have renamed the 'galera' package to
'galera-3'. This patch also handles that change.

These three changes have been combined as without them all together the
gate commit check will not pass.

Closes-Bug: #1430334
Closes-Bug: #1430771
Co-Authored-By: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>
Change-Id: Icc8e34d0984c71e23c5cb261abe8814026cd3f7f