Set container_networks to a dict rather than None
Ensure container_networks is an empty dict rather than None. This
ensures that if there are no container_networks on a host the
dynamic_inventory won't crash. container_networks is a required host
variable, setting it to an empty dict allows it to be ignored when
looping through if there are no container_networks.

Change-Id: Id0517ac6e06e704cf71964c03f016989db51fde5
Closes-Bug: #1430430